Struct oak_restricted_kernel::attestation::AttestationReportData

pub struct AttestationReportData {

    pub version: u32,
    pub guest_svn: u32,
    pub policy: GuestPolicy,
    pub family_id: [u8; 16],
    pub image_id: [u8; 16],
    pub vmpl: u32,
    pub signature_algo: u32,
    pub current_tcb: TcbVersion,
    pub platform_info: u64,
    pub author_key_en: u64,
    pub report_data: [u8; 64],
    pub measurement: [u8; 48],
    pub host_data: [u8; 32],
    pub id_key_digest: [u8; 48],
    pub author_key_digest: [u8; 48],
    pub report_id: [u8; 32],
    pub report_id_ma: [u8; 32],
    pub reported_tcb: TcbVersion,
    pub chip_id: [u8; 64],
    pub committed_tcb: TcbVersion,
    pub current_build: u8,
    pub current_minor: u8,
    pub current_major: u8,
    pub committed_build: u8,
    pub committed_minor: u8,
    pub committed_major: u8,
    pub launch_tcb: TcbVersion,
    /* private fields */

}

The data contained in an attestation report.

See Table 21 in https://www.amd.com/system/files/TechDocs/56860.pdf.

Fields
version: u32

The version of the attestation report format.

This implementation is based on version 2.

guest_svn: u32

The guest security version number.

policy: GuestPolicy

The policy required by the guest VM to be launched.

family_id: [u8; 16]

The family ID provided at launch.

image_id: [u8; 16]

The image ID provided at launch.

vmpl: u32

The VMPL value that was passed in the request.

signature_algo: u32

The algorithm used to sign the report.

Use AttestationReportData::get_signature_algo to try to convert this to a SigningAlgorithm enum.

current_tcb: TcbVersion

The current version of each of the components in the Trusted Computing Base (TCB). This could be different from the committed value during provisional execution when firmware is being updated.

platform_info: u64

Information about the platform.

Use AttestationReportData::get_platform_info to try to convert this to a PlatformInfo biflag representation.

author_key_en: u64

The least significant bit indicates Whether the digest of the author key is included in the report, all other bits are reserved and must be zero.

Use AttestationReportData::get_author_key_en to try to convert this to an AuthorKey enum.

report_data: [u8; 64]

The custom data provided in the attestation request.

measurement: [u8; 48]

The measurement of the VM memory calculated at launch.

host_data: [u8; 32]

Custom data provided by the hypervisor at launch.

id_key_digest: [u8; 48]

The SHA-384 digest of the ID public key used to sign the ID block that was provided in SNP_LAUNCH_FINISH.

author_key_digest: [u8; 48]

The SHA-384 digest of the author public key used to certify the ID key, if the least significant bit of author_key_en is 1, or all zeroes otherwise.

report_id: [u8; 32]

The report ID of this guest.

report_id_ma: [u8; 32]

The report ID of this guest’s migration agent.

reported_tcb: TcbVersion

The reported TCB version that was used to generate the versioned chip endorsement key (VCEK) used to sign this report.

chip_id: [u8; 64]

Identifier unique to the chip, unless the ID has been masked in configuration in which case it is all zeroes.

committed_tcb: TcbVersion

The committed TCB version.

current_build: u8

The build number of the current secure firmware ABI version.

current_minor: u8

The minor number of the current secure firmware ABI version.

current_major: u8

The major number of the current secure firmware ABI version.

committed_build: u8

The build number of the committed secure firmware ABI version.

committed_minor: u8

The minor number of the committed secure firmware ABI version.

committed_major: u8

The major number of the committed secure firmware ABI version.

launch_tcb: TcbVersion

The value of the current TCB version when the guest was launched or imported.

Implementations

impl AttestationReportData

pub fn get_platform_info(&self) -> Option<PlatformInfo>

Gets the platform info field as a PlatformInfo representation if possible.

pub fn get_author_key_en(&self) -> Option<AuthorKey>

Gets the author key enabled field as an AuthorKey enum if possible.

pub fn get_signature_algo(&self) -> Option<SigningAlgorithm>

Gets the signing algorithm field as a SigningAlgorithm enum if possible.

pub fn validate(&self) -> Result<(), &'static str>

Checks that fields with specific expected values or ranges are valid and the reserved bytes are all zero.