Module oak_restricted_kernel::snp_guest

Structs

AttestationReport

A signed attestation report.

AttestationReportData

The data contained in an attestation report.

EcdsaSignature

An ECDSA signature.

GuestPolicy

The required policy for a guest to run.

TcbVersion

The version of all the components in the Trusted Computing Base (TCB).

Enums

AuthorKey

Whether the author key digest is included in the report.

Functions

get_attestation

Requests an attestation report.

get_derived_key

Requests a derived key.

Type Aliases

DerivedKey